fixing up project structure, linters, and tests
added a .gitignore to get rid of ide files that shouldn't be stored.
hooked up an autopep8 file watcher.  This will auto-format python files when they are saved.  This is stricter than pycharm's formatting, as well as stricter than pycodestyle.
added autopep8 as a linter, using diff mode and exit code.
providing docstrings everywhere to satisfy pydocstyle
changing package "main" to "src" which is more standard.