Create a graph class with shortest path method. (#8)
Graphs are constructed from vertices (any Hashable intended to be a vertex label) and edges (pairs of vertices with a weight/distance).  Shortest paths indicate distance from a single initial vertex and provide _all_ available paths of this distance.

This will be a central tool for movement calculation, depending on how the graphs are constructed:
* determine focus and movement by constructing a graph of allowable moves, filtering out any destination that the monster cannot end a move on, calculating from which vertices the monster can attack, and then choosing the shortest path from amongst these candidates.  Treat difficult terrain outgoing edges with weight 2.
* repeating the above but treating negative hexes as movable (traps and hazardous terrain) if there was no focus.
* treating all non-wall hexes as movable for the purpose of calculating proximity.  Hex-geometry will work in most cases but fails to take walls into account.

This leaves line-of-sight calculations but everything else should be largely handled.